+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.7.42] - 2020-06-08 13:35:00
+### Fixed
+- Print price from index
+### Added
+- Remove native layered navigation block in ISP smart navigation context for optimization purposes
+
 ## [4.7.41] - 2020-06-04 13:35:00
 ### Fixed
 - Automatically detect product stores by product ID in save after custom purchase event